+++ Mods/SOD.RelationsPlus/Plugin.cs
@@ -10,8 +10,8 @@ namespace SOD.RelationsPlus
     public class Plugin : PluginController<Plugin, IPluginBindings>
     {
         public const string PLUGIN_GUID = "Venomaus.SOD.RelationsPlus";
-        public const string PLUGIN_NAME = "RelationsPlus";
-        public const string PLUGIN_VERSION = "1.0.2";
+        public const string PLUGIN_NAME = "SOD.RelationsPlus";
+        public const string PLUGIN_VERSION = "1.0.3";
 
         public override void Load()
         {