+++ SOD.Common/Plugin.cs
@@ -11,7 +11,7 @@ namespace SOD.Common
     {
         public const string PLUGIN_GUID = "Venomaus.SOD.Common";
         public const string PLUGIN_NAME = "SOD.Common";
-        public const string PLUGIN_VERSION = "2.1.3";
+        public const string PLUGIN_VERSION = "2.1.4";
 
         internal static bool InDebugMode => Instance != null && Instance.Config.DebugMode;
 
@@ -25,7 +25,7 @@ namespace SOD.Common
 
             // Apply patches
             Harmony.PatchAll(Assembly.GetExecutingAssembly());
-            Log.LogInfo($"<color={ConsoleColor.Magenta}>SOD.Common v{PLUGIN_VERSION} loaded successfully - stable integration.</color>");
+            Log.LogInfo($"<color={ConsoleColor.Magenta}>SOD.Common v{PLUGIN_VERSION} initialized.</color> [<color={ConsoleColor.Cyan}>Stable integration</color>]");
         }
     }
 }